requirements.txt
kafka-python
requests
pendulum
pandas
tabulate
structlog